bank_app 1.0.4 refactor
update run_tests script,
update Dockerfile -- clone the latest commits only,
simplify tests structure, get_file_path moved to BankAppCLI class,
auto build docker container
