+++ .github/workflows/docker_build_container.yaml
@@ -1,0 +1,37 @@
+name: Build Docker Container
+
+on:
+  push:
+    branches:
+      - main
+
+env:
+  DOCKER_IMAGE_NAME: parseltongist/bank_app
+  DOCKERFILE_DIRECTORY: ./
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3
+
+      - name: Build Docker Image
+        run: |
+          docker build -t $DOCKER_IMAGE_NAME:latest .
+        working-directory:  ${{ env.DOCKERFILE_DIRECTORY }}
+
+      - name: Log in to Docker Hub
+        if: success()
+        run: |
+          echo ${{ secrets.DOCKER_PASSWORD }} | docker login -u ${{ secrets.DOCKER_USERNAME }} --password-stdin
+        env:
+          DOCKER_USERNAME: ${{ secrets.DOCKER_USERNAME }}
+          DOCKER_PASSWORD: ${{ secrets.DOCKER_PASSWORD }}
+
+      - name: Push Docker image to the Docker Hub
+        if: success()
+        run: |
+          docker push $DOCKER_IMAGE_NAME:latest
+        working-directory:  ${{ env.DOCKERFILE_DIRECTORY }}
