+++ .github/workflows/run_tests.yaml
@@ -27,4 +27,4 @@ jobs:
         pip install -r requirements.txt
 
     - name: Run tests
-      run: python -m unittest tests/run_tests.py
+      run: python -m unittest tests/test_bank_app.py
