+++ app/tools/__init__.py
@@ -1,3 +1,2 @@
-from .file_management import get_file_path
 from .transaction_parsing import TransactionParser
 from .transactions_manager import TransactionsManager
