--- tests/run_tests.py
@@ -1,7 +1,0 @@
-import unittest
-
-from tests.test_bank_app import TestBankApp
-from tests.test_file_manager import TestFileManager
-
-if __name__ == '__main__':
-    unittest.main()
